// apps/frontend/src/interfaces/product.ts
export interface IProduct {
	readonly id: string[];
	readonly userId: string[];
	readonly username: string[];
	readonly avatar: string[];
	readonly shopName: string[];
	readonly shopId: string[];
	readonly images: string[];
	readonly comments: string[];
	readonly date: string[];
	readonly text: string[];
	readonly likes: string[];
	readonly didLike: string[];
	readonly premium: string[];
}